refactor(package): Docker based development friendly changes
Closes https://github.com/MumbaiHackerspace/Visage/issues/3
