+++ docker-compose.yml
@@ -1,0 +1,20 @@
+version: '3.1'
+
+services:
+  photos:
+    build:
+      context: ./services/photos
+      args:
+        - NODE_ENV=development
+    # you can use legacy debug config or new inspect
+    #command: ../node_modules/.bin/nodemon --debug=0.0.0.0:5858
+    command: ../node_modules/.bin/nodemon --inspect=0.0.0.0:9229
+    ports:
+      - "80:80"
+      - "5858:5858"
+      - "9229:9229"
+    volumes:
+      - ./services/photos:/opt/app
+    environment:
+      - NODE_ENV=development
+
