+++ services/photos/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "react-dropzone-component-example",
+  "name": "Visage Photo Upload",
   "version": "1.0.0",
   "description": "",
-  "main": "build.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack && nodemon server.js",
@@ -21,7 +21,9 @@
     "react-dom": "^15.3.1",
     "react-dropzone-component": "^2.0.0-beta1",
     "react-router-dom": "^4.1.1",
-    "router": "^1.3.1"
+    "router": "^1.3.1",
+    "morgan": "^1.8.1",
+    "project-oxford": "^1.8.0"
   },
   "devDependencies": {
     "babel-core": "^6.14.0",
