+++ services/photos/src-server/ProcessAttendeeImage.js
