+++ services/photos/src/RegisterAttendeeFace.jsx
