+++ services/photos/src/RegisterAttendeeFaceMain.jsx
