+++ services/photos/webpack.config.js
