--- visage-isomorphic/.babelrc
@@ -1,1 +1,0 @@
-{ "presets": ["latest"] }
